Add more examples to the sample queres
added more samples (off the top of my head) that require the llm to provide recipes or ask a follow up. this was done off the top of my head and thus may not be the most diversied set of queries.

however, I attempted to provide different types of requests for different personas (e.g., someone in a hury, a restaurant owner, babysitter, etc.)